fix(lua): correct FFI problem on ARM64
On arriving on my work machine this morning, which is Apple/ARM, I found
this unseemly error:

    NYI: cannot call this C function (yet)

Turns out that this is caused by me going from 8 to 9 parameters in the
`commandt_matcher_new()` function in:

    commit 53a10899a9c862ad4088b95198f6be1e8b67e76e
    Date:   Mon Jul 25 18:50:16 2022 +0200

      feat(lua): implement 'smartcase' functionality

As explained here:

    https://github.com/LuaJIT/LuaJIT/issues/205#issuecomment-236426398

the workaround is to use `uint64_t` or `int64_t` type instead.

+++ lua/wincent/commandt/lib/matcher.h
@@ -32,7 +32,12 @@ matcher_t *commandt_matcher_new(
     bool never_show_dot_files,
     bool recurse,
     bool smart_case,
-    unsigned threads
+
+    // For Apple/ARM64: this would be `unsigned`, but their ABI require 64-bit
+    // type.
+    //
+    // See: https://github.com/LuaJIT/LuaJIT/issues/205#issuecomment-236426398
+    uint64_t threads
 );
 
 /**
